requirements.txt
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.10.0
apturl==0.5.2
asgiref==3.9.1
async-timeout==4.0.1
attrs==21.2.0
beautifulsoup4==4.10.0
blinker==1.9.0
Brlapi==0.8.3
Brotli==1.0.9
cachetools==6.2.0
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.4.3
click==8.0.3
colorama==0.4.4
colorlog==6.9.0
command-not-found==0.3
configobj==5.0.6
crcmod==1.7
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.8
distro==1.7.0
Django==5.2.6
django-embed-video==1.4.10
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
easy-thumbnails==2.10.1
exceptiongroup==1.3.0
eyeD3==0.8.10
filelock==3.14.0
gitdb==4.0.12
GitPython==3.1.45
grpcio==1.30.2
h11==0.16.0
hidapi==0.14.0.post4
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
idna==3.3
ifaddr==0.1.7
IMDbPY==2021.4.18
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.6
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libvirt-python==8.0.0
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==4.0.0
MarkupSafe==2.0.1
mdurl==0.1.2
more-itertools==8.10.0
mutagen==1.45.1
narwhals==2.2.0
nemo-emblems==6.0.1
netaddr==0.8.0
netifaces==0.11.0
numpy==2.2.6
oauthlib==3.2.0
ollama==0.5.3
onboard==1.4.1
packaging==25.0
PAM==0.4.2
pandas==2.3.2
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.2.1
prompt_toolkit==3.0.52
protobuf==6.32.0
psutil==5.9.0
ptyprocess==0.7.0
pyarrow==21.0.0
pycairo==1.20.1
pycryptodomex==3.11.0
pycups==2.0.1
pycurl==7.44.1
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pyelftools==0.27
Pygments==2.19.2
PyGObject==3.42.1
PyICU==2.8.1
pyinotify==0.9.6
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
PyJWT==2.3.0
pylzma==0.5.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyparted==3.11.7
pyperclip==1.10.0
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
python-gnupg==0.4.8
python-magic==0.4.24
python-xlib==0.29
pytz==2022.1
pyusb==1.3.1
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
qrcode==7.3.1
questionary==2.1.1
reportlab==3.6.8
requests==2.32.5
requests-file==1.5.1
rfc5424-logging-handler==1.4.3
rich==14.1.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
setproctitle==1.2.2
six==1.16.0
smbus==1.1.post2
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.3.1
sqlparse==0.5.3
ssh-import-id==5.11
streamlit==1.49.1
systemd-python==234
tenacity==9.1.2
texttable==1.6.4
tinycss2==1.1.1
tldextract==3.1.2
toml==0.10.2
tornado==6.5.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.2
ubuntu-drivers-common==0.0.0
ufw==0.36.1
Unidecode==1.3.3
urllib3==1.26.5
uv==0.8.17
virtualenv==20.26.1
wadllib==1.3.6
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.2.3
websockets==9.1
xdg==5
xkit==0.0.0
xlrd==1.2.0
yara-python==4.5.1
yt-dlp==2022.4.8
zipp==1.0.0
